Add return type annotations to getRank and getCrowdingDistance for type stability
The solution attributes dict is Dict{Any,Any}, so Julia infers the return
type of get() as Any. Adding explicit ::Int and ::Float64 return type
assertions eliminates type ambiguity for callers in the comparator hot path
(sort during replacement), reducing dynamic dispatch overhead.

+++ src/util/densityEstimator.jl
@@ -2,8 +2,8 @@ function maxCrowdingDistanceValue()
     return typemax(Float64)
 end
 
-@inline function getCrowdingDistance(solution::T) where {T<:Solution}
-    return get(solution.attributes, "CROWDING_DISTANCE_ATTRIBUTE", 0.0)
+@inline function getCrowdingDistance(solution::T)::Float64 where {T<:Solution}
+    return get(solution.attributes, "CROWDING_DISTANCE_ATTRIBUTE", 0.0)::Float64
 end
 
 @inline function setCrowdingDistance(solution::T, crowdingDistance::Float64) where {T<:Solution}
